refactoring of node code to move os specific code to different file
